Refact. Change mocks stracture
